move Function to Parameters module
